pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "phone-address-api"
version = "0.1.0"
description = "FastAPI-сервис для работы с контактными данными"
authors = [
    { name = "Andrey Kulenko", email = "andrey@kulenko.su" }
]
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "pydantic-extra-types",
    "pydantic-settings",
    "phonenumbers",
    "redis", 
    "uvicorn"
]

[project.scripts]
phoneaddress = "phoneaddress:main"

[tool.hatch.build.targets.wheel]
packages = ["phoneaddress"]
